requirements.txt
# requirements.txt

# Core requirements
composio-openai>=0.1.0
openai>=0.27.0
litellm>=1.58.0
google-genai>=1.3.0

# Websocket server support
websockets>=12.0

# Env file loading
python-dotenv>=0.21.0

# Replit environment, if needed
replit>=3.2.2

# For Twitter OAuth in skill_x_api.py
requests>=2.28.0
requests_oauthlib>=1.3.1

# HTTP client
aiohttp>=3.8.0

# Data processing
numpy>=1.24.0
pillow>=9.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Development tools
pip-tools>=6.13.0